test: refactor appearance of anchor tag test
This commit refactors the appearance of the anchor tag when a user DOE
NOT apply the href attribute. The previous test was using a expectation
around not using the href and the appearance of the CTA slot. The CTA
slot is present in the DOM regardless of the href or adding content.

Changes to be committed:
modified:   test/auro-card.test.js